Rename class list module file.
