+++ lib/svgo/css-class-list.js
