+++ lib/svgo/svg2js.js
@@ -2,7 +2,7 @@
 
 var SAX = require('sax'),
     JSAPI = require('./jsAPI.js'),
-    CSSClassList = require('./class-list'),
+    CSSClassList = require('./css-class-list'),
     entityDeclaration = /<!ENTITY\s+(\S+)\s+(?:'([^\']+)'|"([^\"]+)")\s*>/g;
 
 var config = {
